requirements.txt
certifi==2020.12.5
mysql-connector-python==8.0.23
protobuf==3.14.0
six==1.15.0
wincertstore==0.2
